Reduce memory usage by making Pipe smaller (#58)
`Pipe` contains a `Kind` which contains a `KindWidth` which is defined as follows:

```rust
#[derive(Clone, Copy)]
enum KindWidth {
    Auto,
    Custom(usize),
}
```

As it stands, the size of `KindWidth` is 16: 8 bytes to represent the `usize`, plus another byte so we can tell whether the variant is `Auto` or `Custom`. Sizes have to be multiples of 8 bytes for alignment, though, so Rust adds another 7 bytes of padding for a total of 16 bytes.

If `KindWidth::Custom` is forbidden to be 0 (we can use `std::num::NonZeroUsize` for this), then Rust can use the ‘niche value optimisation’, where if the value is 0 it knows it must be `KindWidth::Auto`; otherwise, it knows the variant is `KindWidth::Custom`. This leads to `KindWidth`’s size being just 8 bytes, since it doesn’t take any extra data to store which variant it is in.

Therefore, this has reduced the size of `Pipe` from 80 bytes to 72 bytes.

It doesn’t make sense for the width of a pipe kind’s characters to be zero, so we’ve

- [made an illegal state unrepresentable](https://fsharpforfunandprofit.com/posts/designing-with-types-making-illegal-states-unrepresentable/). This prevents us from accidentally introducing a bug by making a pipe kind’s characters have zero width.
- reduced memory usage. Prior to this PR, memory usage when running using `cargo r --release -- -p 929292 -d 0 -r 0` was around 100MB, while with it memory usage is ~90MB.