Always inline Terminal::move_cursor_to (#60)
This is the only `Terminal` function that wasn’t inlined automatically by LLVM, so I thought it would be good to force inlining it. All it does is forward to `core::fmt::write`, so I’d hazard a guess that this makes things faster, especially considering that it takes around 20% of pipes-rs’s runtime.