Don’t unnecessarily store all on-screen text in `Screen` (#86)
Currently all text on-screen is stored in `Screen` so we can correctly calculate what portion of the screen has text on it so we know when to clear the screen. Turns out we don’t ever examine _what_ the text is, just whether it’s present, so this PR just stores a `bool` for each cell instead of a `char`. Technically we could use [bitvec](https://crates.io/crates/bitvec) instead of a `Vec<bool>`, but this would reduce code clarity without a guaranteed performance win (indexing bits is more expensive than indexing bytes).