Update number of covered screen cells incrementally (#118)
Currently we loop over all cells on the screen to determine what portion of them are covered. This feels very wasteful to me, so here I’ve changed the code to keep track of how many cells are covered incrementally as the screen is printed to and cleared. This way determining what portion of the screen is covered is $O(1)$ instead of $O(N)$.